+++ src/components/VestingPage/index.tsx
@@ -65,13 +65,13 @@ const VestingPage = () => {
 
   const isWaitingCycle = useMemo(
     () => vestingInfo?.[activeChainTab]?.lastWithdrawalTime && Date.now() - parseInt(vestingInfo[activeChainTab].lastWithdrawalTime?.toString()) * 1000 < 1641600000,
-    [vestingInfo?.[activeChainTab].lastWithdrawalTime , activeChainTab],
+    [vestingInfo?.[activeChainTab]?.lastWithdrawalTime , activeChainTab],
   );
   const vestingValueAmount = useMemo(
     () => (vestingInfo?.[activeChainTab].vestingValue && activeChainTab
       ? new TokenAmount(HAKKA[activeChainTab || 1], vestingInfo[activeChainTab].vestingValue.toString())
       : new TokenAmount(HAKKA[activeChainTab || 1], '0')),
-    [vestingInfo?.[activeChainTab].vestingValue, activeChainTab],
+    [vestingInfo?.[activeChainTab]?.vestingValue, activeChainTab],
   );
 
   const vestingValuePrice = useMemo(
@@ -82,7 +82,7 @@ const VestingPage = () => {
     () => (vestingInfo?.[activeChainTab].vestingProportion && activeChainTab
       ? new TokenAmount(HAKKA[activeChainTab || 1], vestingInfo[activeChainTab].vestingProportion.toString())
       : new TokenAmount(HAKKA[activeChainTab || 1], '0')),
-    [vestingInfo?.[activeChainTab].vestingProportion, activeChainTab],
+    [vestingInfo?.[activeChainTab]?.vestingProportion, activeChainTab],
   );
 
   const countdownRenderer = ({
