+++ src/hooks/vesting/useVestingInfo.ts
@@ -1,20 +1,7 @@
-import { JSON_RPC_PROVIDER } from './../../constants/index';
-import {
-  Contract as MulticallContract,
-  Provider as MulticallProvider,
-} from '@pelith/ethers-multicall';
-import { useWeb3React } from '@web3-react/core';
-import {
-  ChainDataFetchingState,
-  VESTING_ADDRESSES,
-  ChainId,
-} from '../../constants';
-import throttle from 'lodash/throttle';
+import { useEffect, useState } from 'react';
 import { BigNumber } from '@ethersproject/bignumber';
-import { AddressZero } from '@ethersproject/constants';
-import { useEffect, useMemo, useState } from 'react';
-import { useBlockNumber } from '../../state/application/hooks';
-import VESTING_ABI from '../../constants/abis/vesting.json';
+import { ChainId } from '../../constants';
+import useFetchVestingInfo from './useFetchVestingInfo';
 
 export type VestingInfoType = {
   [chainId in ChainId]?: {vestingValue?: BigNumber, vestingProportion?: BigNumber, lastWithdrawalTime?: BigNumber};
@@ -22,75 +9,24 @@ export type VestingInfoType = {
 
 export default function useVestingInfo(): {
   vestingInfo: VestingInfoType;
-  fetchDataState: ChainDataFetchingState;
 } {
-  const { account } = useWeb3React();
-  const latestBlockNumber = useBlockNumber();
-  const [vestingInfo, setVestingInfo] = useState<
-    VestingInfoType
-  >();
-  const [transactionSuccess, setTransactionSuccess] = useState(false);
+  const [vestingInfo, setVestingInfo] = useState<VestingInfoType>();
 
-  const fetchDataState: ChainDataFetchingState = useMemo(() => {
-    return transactionSuccess
-      ? ChainDataFetchingState.SUCCESS
-      : ChainDataFetchingState.LOADING;
-  }, [transactionSuccess]);
-
-  const providers = JSON_RPC_PROVIDER;
-
-  const getVestingInfo = async (chainId: ChainId, account: string) => {
-    const multicallProvider = new MulticallProvider(
-      providers[chainId],
-      chainId
-    );
-    if (VESTING_ADDRESSES[chainId] === AddressZero)
-      return [chainId, undefined];
-    const vestingContract = new MulticallContract(
-        VESTING_ADDRESSES[chainId],
-        VESTING_ABI
-    );
-    const [vestingValue, vestingProportion, lastWithdrawalTime] = await multicallProvider.all([
-        vestingContract.balanceOf(account),
-        vestingContract.proportion(),
-        vestingContract.lastWithdrawalTime(account),
-    ]);
-    return [chainId, {vestingValue, vestingProportion, lastWithdrawalTime}];
-  };
-
-  const fetchVestingInfo = async (account: string) => {
-    setTransactionSuccess(false);
-    try {
-      const vestingInfoList = [
-        getVestingInfo(ChainId.MAINNET, account),
-        getVestingInfo(ChainId.BSC, account),
-        getVestingInfo(ChainId.POLYGON, account),
-        getVestingInfo(ChainId.FANTOM, account),
-      ];
-      if (process.env.GATSBY_ENV === 'development') {
-        vestingInfoList.push(
-          getVestingInfo(ChainId.KOVAN, account),
-        );
-      }
-
-      const vestingInfoResults = await Promise.all(vestingInfoList);
-      setVestingInfo(Object.fromEntries(vestingInfoResults));
-      setTransactionSuccess(true);
-    } catch (e) {
-      console.log(e);
-      console.log('fetch user vesting info error');
-    }
-  };
-
-  const throttledFetchVestingInfo = useMemo(
-    () => throttle(fetchVestingInfo, 2000),
-    []
-  );
+  const {fetchVestingInfoResult: mainnetVestingInfo} = useFetchVestingInfo(ChainId.MAINNET);
+  const {fetchVestingInfoResult: bscVestingInfo} = useFetchVestingInfo(ChainId.BSC);
+  const {fetchVestingInfoResult: fantomVestingInfo} = useFetchVestingInfo(ChainId.FANTOM);
+  const {fetchVestingInfoResult: polygonVestingInfo} = useFetchVestingInfo(ChainId.POLYGON);
+  const {fetchVestingInfoResult: kovanVestingInfo} = useFetchVestingInfo(ChainId.KOVAN);
 
   useEffect(() => {
-    if (account === AddressZero || !account) return;
-    throttledFetchVestingInfo(account);
-  }, [latestBlockNumber, account]);
+    setVestingInfo(Object.fromEntries([
+      mainnetVestingInfo || [[ChainId.MAINNET], undefined], 
+      bscVestingInfo || [[ChainId.BSC], undefined], 
+      fantomVestingInfo || [[ChainId.FANTOM], undefined], 
+      polygonVestingInfo || [[ChainId.POLYGON], undefined], 
+      kovanVestingInfo || [[ChainId.KOVAN], undefined]
+    ]));
+  }, [mainnetVestingInfo, bscVestingInfo, fantomVestingInfo, polygonVestingInfo, kovanVestingInfo]);
 
-  return { vestingInfo, fetchDataState };
+  return { vestingInfo };
 }
